[util] Add possibility to run 1-time benchmarks
We can have better control over benchmarking
The goal is to produce files that are the same as benchmarks so we can use `benchstat`.
b.N is not configurable so the problem is on tight control of benchmarks, we cannot have a modular b.N.
So we need a way to generate constant-length benchmarks

 bit of a hacky way to run only-once (b.N=1) benchmarks through tests (so we don't compile
benchmarks into the main library as well as they wouldn't appear on the doc site).
We cannot create a separate package because hashDisk is an internal struct (but we still want to
benchmark it).
Usage:
EXTENDED_BENCH_FILE env var must be set to a path where we will write the results
You can then call all the tests that starts with TestExtendedBench
Example:
EXTENDED_BENCH_FILE=/tmp/results go test -run TestExtendedBench